share hook for query and mutation in edit view
